requirements.txt
flask
flask-socketio
flask-login
python-socketio
Flask-Babel
flask-migrate
raven
pyyaml
netifaces
docopt
lxml
celery
Flask-Celery-Tools
Flask-SQLAlchemy
eventlet
psutil
pyserial
setuptools
python-pam
flask-cors
flask-jwt-extended~=3.24.1